backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Azure SDK (for production integration)
# Uncomment when ready to integrate with Azure services
# azure-identity==1.15.0
# azure-cosmos==4.5.1
# azure-search-documents==11.4.0
# openai==1.10.0  # Azure OpenAI support

# HTTP client (for external API calls)
httpx==0.26.0

# Utilities
pydantic-settings==2.1.0

# Development dependencies
# pytest==7.4.4
# pytest-asyncio==0.23.3
# black==24.1.1
# flake8==7.0.0
